add conftest.py to pull test input data
This hook will run by pytest one time before execution all the particular
tests. It does checkout the needed 'extern/test-data' directory from the
separate branch test-data'.

'extern/test-data' is locatated in the toplevel directory of our git
repository.

We are doing this to keep this huge amount of data outside of
the git history of our master branch.

The data will be transparently pulled in, when running the tests in
the master branch. This will be done either by

  - some hook magic via conftest.py (that's what this commit does now)
  - or via git submodule (can be added later)

The midterm goal would be to get rid off this separate branch again, as
it might be a bit complicated for developers to keep the test-data
and actual tests in the master branch in sync.

Steps needed to remove the separate branch again would be:

 - Make the input data smaller in general. Find smaller input data that
   still provide a good test coverage, also for corner cases.
 - Use text input data (.csv or .py) instead of binary (.pkl). This can
   be much better compressed by git. Also such text data can be easier
   reviewed in pull requests, or later changed/extended, without
   affecting the size of .git badly (diff compression).
 - Re-use/share large input data among several tests, Avoid adding
   too many large files at all.