[REF] Updates context.genetics function
As of abagen v0.1 the main get_expression_data workflow supports
surface atlases (in a variety of spaces; default: fsaverage5). This
removes the surface -> volume projection from the context.genetics
module and simply passes the surface parcellation directly to
abagen.

Also, a few new parameters were added / updated, and the minimum
abagen version was pinned to 0.1.