[REF] Replaces numpy_groupies with np.bincount (#139)
* [REF] Removes numpy_groupies

Replaces all instances of numpy_groupies.aggregate with equivalent
calls to np.bincount. Removes dependency for numpy_groupies

* [REF] Suppress warning in data.mesh_smooth

In tests (and, presumably, in some real cases scenarios) smooth_mesh
raises a warning due to a divide-by-zero error. This just wraps the
offending line in a context manager to ignore the warning